+++ x/crosschain/keeper/msg_server_add_to_outtx_tracker.go
@@ -24,7 +24,7 @@ import (
 // Authorized: admin policy group 1, observer.
 func (k msgServer) AddToOutTxTracker(goCtx context.Context, msg *types.MsgAddToOutTxTracker) (*types.MsgAddToOutTxTrackerResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
-	chain := k.zetaObserverKeeper.GetParams(ctx).GetChainFromChainID(msg.ChainId)
+	chain := k.zetaObserverKeeper.GetSupportedChainFromChainID(ctx, msg.ChainId)
 	if chain == nil {
 		return nil, observertypes.ErrSupportedChains
 	}
@@ -47,7 +47,7 @@ func (k msgServer) AddToOutTxTracker(goCtx context.Context, msg *types.MsgAddToO
 	if msg.Proof == nil { // without proof, only certain accounts can send this message
 		adminPolicyAccount := k.zetaObserverKeeper.GetParams(ctx).GetAdminPolicyAccount(observertypes.Policy_Type_group1)
 		isAdmin := msg.Creator == adminPolicyAccount
-		isObserver := k.zetaObserverKeeper.IsAuthorized(ctx, msg.Creator, chain)
+		isObserver := k.zetaObserverKeeper.IsAuthorized(ctx, msg.Creator)
 
 		// Sender needs to be either the admin policy account or an observer
 		if !(isAdmin || isObserver) {