+++ x/observer/types/params.go
@@ -77,6 +77,7 @@ func (p Params) String() string {
 	return string(out)
 }
 
+// Deprecated: observer params are now stored in core params
 func validateVotingThresholds(i interface{}) error {
 	v, ok := i.([]*ObserverParams)
 	if !ok {
@@ -113,69 +114,4 @@ func (p Params) GetAdminPolicyAccount(policyType Policy_Type) string {
 		}
 	}
 	return ""
-}
-func (p Params) GetParamsForChain(chain *common.Chain) ObserverParams {
-	for _, ObserverParam := range p.GetObserverParams() {
-		if ObserverParam.Chain.IsEqual(*chain) {
-			return *ObserverParam
-		}
-	}
-	return ObserverParams{}
-}
-
-func (p Params) GetParamsForChainID(chainID int64) ObserverParams {
-	for _, ObserverParam := range p.GetObserverParams() {
-		if ObserverParam.Chain.ChainId == chainID {
-			return *ObserverParam
-		}
-	}
-	return ObserverParams{}
-}
-
-func (p Params) GetSupportedChains() (chains []*common.Chain) {
-	for _, observerParam := range p.GetObserverParams() {
-		if observerParam.IsSupported {
-			chains = append(chains, observerParam.Chain)
-		}
-	}
-	return
-}
-
-func (p Params) GetChainFromChainID(chainID int64) *common.Chain {
-	chainList := p.GetObserverParams()
-	for _, observerParam := range chainList {
-		if observerParam.Chain.ChainId == chainID && observerParam.IsSupported {
-			return observerParam.Chain
-		}
-	}
-	return nil
-}
-
-func (p Params) GetChainFromChainName(name common.ChainName) *common.Chain {
-	for _, observerParam := range p.GetObserverParams() {
-		if observerParam.Chain.ChainName == name && observerParam.IsSupported {
-			return observerParam.Chain
-		}
-	}
-	return nil
-}
-
-func (p Params) IsChainSupported(checkChain common.Chain) bool {
-	chains := p.GetSupportedChains()
-	for _, chain := range chains {
-		if checkChain.IsEqual(*chain) {
-			return true
-		}
-	}
-	return false
-}
-
-func (p Params) IsChainIDSupported(checkChainID int64) bool {
-	chains := p.GetSupportedChains()
-	for _, chain := range chains {
-		if chain.ChainId == checkChainID {
-			return true
-		}
-	}
-	return false
 }